UPDATED code in order to follow Spring-Security guidelines in preparation of adding the code to Spring-Security 5. ADDED indices to the MongoAcl domain object in order to improve lookup t imes
This commit also contains a fix for the lookup of children which was not correct
the way it was implemented. Due to the inability of left-joining and filtering
stuff similar to SQL in a MongoDB first each ACL for the provided domain object
is retrieved and afterwards a lookup for ACLs that specify that ACL as parent
is performed and results converted back to ObjectIdentity's.

Unfortunately 2 tests currently fail due to a StackOverflowException caused by
hashCode of AclImpl and AccessControlEntryImpl that are invoking each other
(Spring Secuirty ACL  in the master branch (5.0.1.M1)). Former 4.3.2 version
works without any issue.

TODO: check if the general lookup query can be improved further by using
either a paginated approach or a streaming one